requirements.txt
colorama==0.4.6
fitz==0.0.1.dev2
loguru==0.7.2
Pillow==10.1.0
schedule==1.2.1
toml==0.10.2
tqdm==4.66.1
